Added Resources, Resource handles, and resource managers.
Resources are classes who are managed by a resource manager.
A resource manager is responsible for managing the creation storage and deletion of resources through reference counting.

A resource handle is what the client will receive, the handle calls the resource managers reference counting functions.

ResourceHandle is the base handle class that contains the functionality.
Handle is the templated version that stores a typed resource.