Fleshed out Window class among other things
The window class is now fleshed out with most glfw callbacks added. Users can also hook into glfw callbacks. The class now uses a descriptor for initialisation. This is mirrored in the instance class.

The Wrapper class was added for implicitly converting a class into another type. This is used in the window class to implicitly convert to GLFWwindow*

Logging now uses cleaner code.
The Message class had its message fixed.
The Resource class now use a pointer to a ResourceManager, and ResourceHandles can now be moved and copied.
The ResourceManager class no longer logs every reference and resource created and destroyed.